backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.1
httpx>=0.27.0
Unidecode>=1.3.8
SQLAlchemy>=2.0.35
alembic>=1.13.3
psycopg[binary]>=3.2.3
pytest>=8.3.0
